@@ -11,17 +11,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.2.0",
+    "@angular/animations": "^18.2.12",
     "@angular/cdk": "^18.2.6",
-    "@angular/common": "^17.2.0",
-    "@angular/compiler": "^17.2.0",
-    "@angular/core": "^17.2.0",
-    "@angular/forms": "^17.2.0",
-    "@angular/platform-browser": "^17.2.0",
-    "@angular/platform-browser-dynamic": "^17.2.0",
+    "@angular/common": "^18.2.12",
+    "@angular/compiler": "^18.2.12",
+    "@angular/core": "^18.2.12",
+    "@angular/forms": "^18.2.12",
+    "@angular/platform-browser": "^18.2.12",
+    "@angular/platform-browser-dynamic": "^18.2.12",
     "@angular/pwa": "^17.3.1",
-    "@angular/router": "^17.2.0",
-    "@angular/service-worker": "^17.3.1",
+    "@angular/router": "^18.2.12",
+    "@angular/service-worker": "^18.2.12",
+    "@ctrl/tinycolor": "^4.1.0",
+    "@kolkov/angular-editor": "^3.0.0-beta.0",
+    "@kylepairish/angular-editor": "^3.0.2-beta.0",
     "ajv": "^8.17.1",
     "cron-parser": "^4.9.0",
     "d3": "^7.8.5",
@@ -35,16 +38,16 @@
     "ng-zorro-antd": "^17.2.0",
     "ngx-image-compress": "^15.1.6",
     "ngx-print": "^1.5.1",
-    "pdfmake": "^0.2.10",
+    "pdfmake": "^0.2.15",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "xlsx": "^0.18.5",
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
-    "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.0",
+    "@angular-devkit/build-angular": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.12",
     "@types/d3": "^7.4.3",
     "@types/d3-zoom": "^3.0.8",
     "@types/file-saver": "^2.0.7",
@@ -57,6 +60,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.3.2"
+    "typescript": "~5.5.4"
   }
 }
