@@ -26,7 +26,7 @@ export class EmployeeContractComponent implements OnInit {
       id: [0, []],
       employee: [0, [Validators.required]],
       position: ['', [Validators.required]],
-      // isHourlyRateHourly: [false],
+      isHourlyRateHourly: [false],
       salary: [0, [Validators.required]],
       // overtime: [false],
       // siginout_required: [false],
@@ -46,7 +46,7 @@ export class EmployeeContractComponent implements OnInit {
         id: [this.editableItem.id],
         employee: [this.editableItem.employee, [Validators.required]],
         position: [this.editableItem.position, [Validators.required]],
-        // isHourlyRateHourly: [this.editableItem.isHourlyRateHourly],
+        isHourlyRateHourly: [this.editableItem.isHourlyRateHourly],
         salary: [this.editableItem.salary, [Validators.required]],
         // overtime: [this.editableItem.overtime],
         // siginout_required: [this.editableItem.siginout_required],
@@ -72,7 +72,7 @@ export class EmployeeContractComponent implements OnInit {
         // supervisor: this.employeeForm.value.supervisor,
         salary: this.employeeForm.value.salary,
         // overtime: this.employeeForm.value.overtime,
-        // isHourlyRateHourly: this.employeeForm.value.isHourlyRateHourly,
+        isHourlyRateHourly: this.employeeForm.value.isHourlyRateHourly,
         // workingHours: this.employeeForm.value.workingHours,
         // siginout_required: this.employeeForm.value.siginout_required,
         details: this.employeeForm.value.details,
@@ -84,7 +84,7 @@ export class EmployeeContractComponent implements OnInit {
         // supervisor: this.employeeForm.value.supervisor,
         salary: this.employeeForm.value.salary,
         // overtime: this.employeeForm.value.overtime,
-        // isHourlyRateHourly: this.employeeForm.value.isHourlyRateHourly,
+        isHourlyRateHourly: this.employeeForm.value.isHourlyRateHourly,
         // workingHours: this.employeeForm.value.workingHours,
         // siginout_required: this.employeeForm.value.siginout_required,
         details: this.employeeForm.value.details,
