@@ -7,16 +7,17 @@ export const AppPermissions = {
     ViewPermissions: 5,
     UpdatePermissions: 6,
     ViewTasks: 8,
-    AddUpdateTask: 9,
-    ViewProjectBoard: 10,
+    AddTask: 9,
+    UpdateTask: 10,
+    ViewProjectBoard: 11,
     UpdateClientContract: 13,
     UpdateClientEmployee: 14,
     ViewClientEmployee: 15,
-    AddUpdateTaskWorklog: 16,
+    AddTaskWorklog: 16,
     AddUpdatePayments: 17,
-    AddClients: 11,
-    ViewClients: 12,
-    SwitchClients: 19,
+    AddClients: 18,
+    ViewClients: 19,
+    SwitchClients: 20,
 };
 
 export const PermissionName = {
@@ -27,14 +28,15 @@ export const PermissionName = {
     [AppPermissions.ViewPermissions]: "View Permissions",
     [AppPermissions.UpdatePermissions]: "Manage Permissions",
     [AppPermissions.ViewTasks]: "View Tasks",
-    [AppPermissions.AddUpdateTask]: "Add or Update Task",
+    [AppPermissions.AddTask]: "Add Task",
+    [AppPermissions.UpdateTask]: "Upate Task",
     [AppPermissions.ViewProjectBoard]: "Project Board View",
     [AppPermissions.AddClients]: "Add Clients",
     [AppPermissions.ViewClients]: "View Clients",
     [AppPermissions.UpdateClientContract]: "Update Client Contract",
     [AppPermissions.UpdateClientEmployee]: "Update Client Employee",
     [AppPermissions.ViewClientEmployee]: "View Client Employee",
-    [AppPermissions.AddUpdateTaskWorklog]: "Update Task Worklog",
+    [AppPermissions.AddTaskWorklog]: "Update Task Worklog",
     [AppPermissions.AddUpdatePayments]: "Manage Employee Hourly Rate",
     [AppPermissions.AddUpdatePayments]: "Add or Update Payments",
     [AppPermissions.SwitchClients]: "Switch Clients",
@@ -47,11 +49,13 @@ export const DefaultPermissions = {
     [AppPermissions.UpdateUsers]: false,
     [AppPermissions.AddClients]: false,
     [AppPermissions.UpdatePermissions]: false,
-    [AppPermissions.AddUpdateTask]: false,
+    [AppPermissions.AddTask]: false,
+    [AppPermissions.UpdateTask]: false,
+    
     [AppPermissions.UpdateClientContract]: false,
     [AppPermissions.ViewTasks]: false,
     [AppPermissions.UpdateClientEmployee]: false,
-    [AppPermissions.AddUpdateTaskWorklog]: false,
+    [AppPermissions.AddTaskWorklog]: false,
     [AppPermissions.AddUpdatePayments]: false,
     [AppPermissions.ViewPermissions]: false,
     [AppPermissions.SwitchClients]: false,
